strava_agent/requirements.txt
google-genai
python-dotenv
pytesseract>=0.3.10
Pillow>=10.0.0
pytest>=7.0.0
